move lots of stuff around
